24-07-29 10:00 ~ 24-07-30 09:00  merge

See merge request s11-webmobile1-sub2/S11P12E107!94

+++ backend/sayif/src/main/java/com/ssafy/sayif/team/service/TeamService.java
@@ -24,7 +24,7 @@ public class TeamService {
     private final SimpMessagingTemplate messagingTemplate;
     private final MemberRepository memberRepository;
 
-    public List<Member> getMembersByTeamId(Long teamId) {
+    public List<Member> getMembersByTeamId(Integer teamId) {
         return memberRepository.findByTeamId(teamId);
     }
